Add new PrintTime overload
This overload was added to the IDateTimeFormatter interface in KSP
1.3.1. Doesn't seem to be a documented change, unfortunately, but the
code definitely compiles for KSP 1.3.0 and doesn't for KSP 1.3.1.

I have no idea what the new parameter (bool logEnglish) is supposed to
indicate, so I chose the easy way out and forwarded the call to the
existing PrintTime method, ignoring the new parameter.